DI by hand for all exercise related viewmodels
